% Generated by build script
\name{mt_balance}
\alias{mt_balance}
\title{Balance sheet metrics}
\usage{
mt_balance()
}
\arguments{
\item{ticker}{Stock ticker}
\item{slug}{Company name slug}
\item{period}{"annual", "quarterly", or "both"}
}
\value{tibble: stacked balance sheet metrics}
\description{Balance sheet metrics}
