% Generated by build script
\name{mt_cashflow}
\alias{mt_cashflow}
\title{Cash flow metrics}
\usage{
mt_cashflow()
}
\arguments{
\item{ticker}{Stock ticker}
\item{slug}{Company name slug}
\item{period}{"annual", "quarterly", or "both"}
}
\value{tibble: stacked cash flow metrics}
\description{Cash flow metrics}
