% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt.R
\name{mt_margins}
\alias{mt_margins}
\title{Profitability margins}
\usage{
mt_margins(ticker, slug = NULL, period = "annual")
}
\arguments{
\item{ticker}{Stock ticker}

\item{slug}{Company name slug}

\item{period}{"annual", "quarterly", or "both"}
}
\value{
tibble: stacked margin metrics
}
\description{
Profitability margins
}
