% Generated by build script
\name{mt_margins}
\alias{mt_margins}
\title{Profitability margins}
\usage{
mt_margins()
}
\arguments{
\item{ticker}{Stock ticker}
\item{slug}{Company name slug}
\item{period}{"annual", "quarterly", or "both"}
}
\value{tibble: stacked margin metrics}
\description{Profitability margins}
