% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt.R
\name{mt_valuation}
\alias{mt_valuation}
\title{Valuation ratios (PE, PB, EV/EBITDA)}
\usage{
mt_valuation(ticker, slug = NULL, period = "annual")
}
\arguments{
\item{ticker}{Stock ticker}

\item{slug}{Company name slug}

\item{period}{"annual" or "quarterly"}
}
\value{
tibble: stacked valuation metrics
}
\description{
Valuation ratios (PE, PB, EV/EBITDA)
}
